move handler code from domain to api lib
