+++ src/backend/api/Fusion.Resources.Domain/Fusion.Resources.Domain.csproj
@@ -6,7 +6,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <PackageReference Include="Fusion.Integration.Notification" Version="5.1.0" />
+    <PackageReference Include="AdaptiveCards" Version="2.4.0" />
     <PackageReference Include="MediatR" Version="9.0.0" />
     <PackageReference Include="MediatR.Extensions.Microsoft.DependencyInjection" Version="9.0.0" />
     <PackageReference Include="FluentValidation.AspNetCore" Version="9.5.1" />
